Decoupled Marionette.Callbacks depency on jQuery.Deffered
Setting Marionette.compatMode based on check against $ version string.
If major is less than 2 and minor is less than 5 Marionette.Callbacks
will now return warning messages instead of failing. Obviously the user
of this version will not be able to safely use Marionette.Callbacks so
the check is really there to allow for forwards compatibility in a
project that currently uses less than jQuery 1.5 but may at an
indeterminate time in the future be upgraded to a more recent version.
In short, this keeps it from breaking now and gives the developer/user
a warning message indicating when $.Deffered isn't supported.